backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
PyJWT==2.8.0
python-dotenv==1.0.1
google-generativeai==0.5.2
werkzeug==3.0.3
reportlab==4.2.0
google-search-results==2.4.2
APScheduler==3.10.4
python-pptx==0.6.23
bcrypt==4.1.2
Flask-Limiter==3.5.1
bleach==6.1.0
